+++ packages/apollo-cache-inmemory/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change log
 
 ### vNEXT
+- Added `getCacheKey` function to cacheResolver context [#2998](https://github.com/apollographql/apollo-client/pull/2998)
+
+### 1.1.8
+- dependency updates
 - Fix IntrospectionResultData type definition [#2959](https://github.com/apollographql/apollo-client/issues/2959)
 
 ### 1.1.7