+++ packages/apollo-utilities/CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ### vNext
 
+### 1.1.3
+- dependency updates
+- Added getDirectivesFromDocument utility function
+[PR#2974](https://github.com/apollographql/apollo-client/pull/2974)
+
 ### 1.1.2
 - Add client, rest, and export directives to list of known directives [PR#2949](https://github.com/apollographql/apollo-client/pull/2949)