+++ packages/graphql-anywhere/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change log
 
 ## vNEXT
+
+### 4.1.4
+- dependency updates
 - support null literal
 
 ### 4.1.3